+++ Sources/DangerFixtures/DangerDSLResources/DangerDSLGitLab.swift
@@ -83,6 +83,16 @@ public let DSLGitLabJSON = """
             "web_url": "https://gitlab.com/orta"
           }
         ],
+        "reviewers": [
+          {
+            "id": 377669,
+            "name": "Orta",
+            "username": "orta",
+            "state": "active",
+            "avatar_url": "https://secure.gravatar.com/avatar/f116cb3be23153ec08b94e8bd4dbcfeb?s=80&d=identicon",
+            "web_url": "https://gitlab.com/orta"
+          }
+        ],
         "source_project_id": 10132593,
         "target_project_id": 1620437,
         "labels": [],